requirements.txt
flask
transformers
langid
openai
pydub
requests
flask_cors
python-dotenv
whisper
jinja2
vertexai
google-cloud-translate
transformers
ffmpeg-python
gunicorn
torch